#1, #2, #3
Not sure if this is my long-term implementation, but it's something I can start prototyping with. I think I'll use this style to try and work out a first chapter, and I'll add in each of my mainly desired features along the way. Once there's a fully functional version for the first chapter, I'll begin extracting things into interfaces and try to settle on a long-term design.